Refactor: Split database schema to separate files
